Working on refactoring configuration_unit to build_unit.
