+++ src/crow/unit/CMakeLists.txt
@@ -1,5 +1,5 @@
 # Add source files
 target_sources(${TARGET_CROW_LIB} PRIVATE
-  configuration_unit.cpp
+  build_unit.cpp
   translation_unit.cpp
 )
